did challenge number 3 (somehow my package.json got deleted in the mean time?)
