+++ challenges/arrayBinarySearch/README.md
@@ -1,0 +1,11 @@
+# Search for an Element in an Array
+<!-- Short summary or background information -->
+
+## Challenge
+This code challenge asks that we write a function which returns the index of a desired element in an array. No default array functions can be used. 
+
+## Approach & Efficiency
+The approach I took for this problem was to keep the solution simple. I made a variable called index, which will eventually be returned. The function uses a for loop to go through the array and test equality of the given number and the current element. If they are equal, the index takes on the value of the for loop idx variable, i. Then index is returned. 
+
+## UML Diagram
+<img src="./assets/array-binary-search-whiteboarding.png">
