+++ docs/config/jsdoc.config.json
@@ -1,0 +1,9 @@
+{
+  "source": {
+    "include" : ["./src"]
+  },
+  "opts": {
+    "destination": "./docs/",
+    "recurse": true
+  }
+}
